moved parser classes in dedicated package
