+++ src/main/java/org/jqassistant/plugin/openapi/impl/OpenAPIScannerPlugin.java
@@ -11,6 +11,7 @@ import com.buschmais.jqassistant.plugin.common.api.scanner.filesystem.FileResour
 import io.swagger.v3.oas.models.*;
 import io.swagger.v3.parser.OpenAPIV3Parser;
 import org.jqassistant.plugin.openapi.api.model.*;
+import org.jqassistant.plugin.openapi.impl.parsers.ContractParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
